Move contract update validation tests to a separate file
